user authentication

+++ main_app/routes.py
@@ -1,8 +1,9 @@
-from flask import Blueprint, request, redirect, render_template, url_for, flash
+from flask import Blueprint, request, redirect, render_template, url_for, flash, session
 from flask_pymongo import PyMongo
 from bson.objectid import ObjectId
 from werkzeug.utils import secure_filename
 from markdown import markdown
+from passlib.hash import pbkdf2_sha256
 from main_app.forms import *
 from main_app.models import *
 from main_app import app, db
@@ -13,6 +14,25 @@ main = Blueprint("main", __name__)
 ############################################################
 # HELPER FUNCTIONS
 ############################################################
+
+def hash_password(password):
+    return pbkdf2_sha256.hash(password)
+
+def verify_hash(a, b):
+    return pbkdf2_sha256.verify(a, b)
+
+# TODO: gotta be a better way to handle this
+def check_logged_in():
+    if not "user_id" in session:
+        flash("You must be logged in to do this.")
+        return redirect(url_for("main.login"))
+    return True
+
+def check_correct_login(user_id):
+    if session["user_id"] != user_id:
+        flash("You are not logged in as this user!")
+        return redirect(url_for("main.homepage"))
+    return True
 
 def doc_from_id(collection, id):
     return collection.find_one({"_id": ObjectId(id)})
@@ -42,32 +62,40 @@ def homepage():
 @main.route("/users")
 def users():
     """Display users"""
+    if not (log := check_logged_in()) == True:
+        return log
     users_data = db.users.find()
     return render_template("users.html", users=users_data)
 
 @main.route("/posts")
 def posts():
-    # TODO: jinja template
     """Display posts"""
+    if not (log := check_logged_in()) == True:
+        return log
     posts_data = db.posts.find()
     return render_template("posts.html", posts=posts_data)
 
 @main.route("/users/<user_id>")
 def user(user_id):
     """Display user information"""
+    if not (log := check_logged_in()) == True:
+        return log
     user_data = doc_from_id(db.users, user_id)
     return render_template("user.html", user=user_data)
 
 @main.route("/profiles/<user_id>")
 def profile(user_id):
     """Display profile information"""
+    if not (log := check_logged_in()) == True:
+        return log
     profile_data = db.profiles.find_one({"user_id": user_id})
     return render_template("profile.html", profile=profile_data)
 
 @main.route("/posts/<post_id>", methods=["GET", "POST"])
 def post(post_id):
-    # TODO: jinja template
     """Display post information"""
+    if not (log := check_logged_in()) == True:
+        return log
     post_data = doc_from_id(db.posts, post_id)
     post_data["user"] = doc_from_id(db.users, post_data["user_id"])
     post_data["body"] = markdown(post_data["body"])
@@ -79,7 +107,7 @@ def post(post_id):
     form = ReplyForm("New Reply", "Please fill out reply info:", users)
     if request.method == "POST":
         reply_variables = {
-            "user_id": request.form.get("user_id"),
+            "user_id": session["user_id"],
             "post_id": post_id,
             "body": request.form.get("body")
         }
@@ -92,9 +120,13 @@ def post(post_id):
 @main.route("/create", methods=["GET", "POST"])
 def create_user():
     """User creation"""
+    if "user_id" in session:
+        flash("You are already logged in!")
+        return redirect(url_for("main.homepage"))
     form = UserForm("Create User", "Please fill out your info:")
     if request.method == "POST":
         new_user = blank_user(**request.form)
+        new_user["password"] = hash_password(new_user["password"])
         if db.users.count_documents({ "username": request.form.get("username")}, limit=1):
             flash("Username already exists!")
             return redirect(url_for("main.create_user"))
@@ -103,6 +135,7 @@ def create_user():
         new_user_email = new_user["email"]
         new_profile = blank_profile(str(new_user_id), new_user_email)
         db.profiles.insert_one(new_profile)
+        session["user_id"] = str(new_user_id)
         flash("User created successfully.")
         return redirect(url_for("main.user", user_id=new_user_id))
     return render_template("form.html", form=form)
@@ -110,10 +143,12 @@ def create_user():
 @main.route("/create-post", methods=["GET", "POST"])
 def create_post():
     """Post creation"""
+    if not (log := check_logged_in()) == True:
+        return log
     users = db.users.find()
     form = PostForm("Create Post", "Please fill out post info:", users)
     if request.method == "POST":
-        new_post = blank_post(**request.form)
+        new_post = blank_post(user_id=session["user_id"], **request.form)
         db.posts.insert_one(new_post)
         flash("Post created successfully.")
         return redirect(url_for("main.homepage"))
@@ -121,6 +156,10 @@ def create_post():
 
 @main.route("/edit-user/<user_id>", methods=["GET", "POST"])
 def edit_user(user_id):
+    if not (log := check_logged_in()) == True:
+        return log
+    if not (log := check_correct_login(user_id)) == True:
+        return log
     user_data = doc_from_id(db.users, user_id)
     form = UserForm("Edit User", "Please edit your info:")
     form.set_values(user_data)
@@ -133,6 +172,10 @@ def edit_user(user_id):
 
 @main.route("/edit-profile/<user_id>", methods=["GET", "POST"])
 def edit_profile(user_id):
+    if not (log := check_logged_in()) == True:
+        return log
+    if not (log := check_correct_login(user_id)) == True:
+        return log
     profile_data = db.profiles.find_one({"user_id": user_id})
     form = ProfileForm("Edit Profile", "Please edit your info:")
     form.set_values(profile_data)
@@ -151,11 +194,12 @@ def edit_profile(user_id):
 
 @main.route("/delete/<user_id>", methods=["GET", "POST"])
 def delete_user(user_id):
-    form = DeleteUserForm()
+    if not (log := check_correct_login(user_id)) == True:
+        return log
     user_data = doc_from_id(db.users, user_id)
+    form = DeleteUserForm()
     if request.method == "POST":
-        password = request.form.get("password")
-        if user_data["password"] == password:
+        if verify_hash(request.form.get("password"), user_data["password"]):
             db.users.delete_one(user_data)
             db.profiles.delete_one({"user_id": user_id})
             flash("User deleted successfully.")
@@ -169,10 +213,18 @@ def login():
     form = LoginForm("Log In", "Please enter your credentials:")
     if request.method == "POST":
         username = request.form.get("username")
-        password = request.form.get("password")
-        for db_password in db.users.find({"username": username}):
-            if password == db_password["password"]:
+        for document in db.users.find({"username": username}):
+            if verify_hash(request.form.get("password"), document["password"]):
+                session["user_id"] = str(document["_id"])
                 flash("Password correct!")
                 return redirect(url_for("main.homepage"))
         flash("Password incorrect. Please try again.")
     return render_template("form.html", form=form)
+
+@main.route("/logout")
+def logout():
+    if "user_id" in session:
+        session.pop("user_id")
+    else:
+        flash("You are not logged in.")
+    return redirect(url_for("main.homepage"))
